+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Fixed
+
+## [1.3.1] - 2022-02-11
+
+### Fixed
+
 - add TargetGroup, GroupPlace to [Group-Information](https://github.com/5pm-HDH/churchtools-api/pull/64)
 - generate Registration-Link in [PublicGroup](https://github.com/5pm-HDH/churchtools-api/pull/64)